Add Codable support to User class manually
This was implemented via three additions: first, an enumerated type
that conforms to CodingKey protocol, where each case is data we want
to save/load; second, the required custom initializer that takes an
instance of Decoder; and finally, an encoding function. This enables
saving and loading data in a custom manner.